Added i8, i6 and i32.
Deleted the documentation.
Enabled all warnings for guest gcc.
Addded error for to big integers.

+++ tests/valid/functions/hello_world.rs
@@ -2,7 +2,7 @@
 output: "Hello, World!\n"
 */
 
-[extern] putchar(c: i64): i64;
+[extern] putchar(c: i32): i32;
 
 main(): i32 {
     putchar(72);
